+++ src/api/ruma_wrapper/mod.rs
@@ -1,4 +1,5 @@
-pub(crate) mod axum;
+mod auth;
+mod request;
 mod xmatrix;
 
 use std::ops::Deref;
